Feat: Add breakpoints to the placeholder page styling
It became clear that simply relying on a fluid width layout wasn't
going to work for the logo and title. This was especially true when
wanting to keep the title text width the same as the logo width.
The use of breakpoints will resolve this problem. I have started
the process of styling the responsive layout of the placeholder
page. So far, I have only made the body width and margins
responsive.